Spread by CPUs

+++ 4-2_Dezentral/src/edu/hm/cs/vss/Client.java
@@ -395,7 +395,7 @@ public class Client extends UnicastRemoteObject implements IClient {
 	}
 	
 	@Override
-	public List<PhilosopherBackup> exportPhilosophers(int nrPhilosophers)
+	public List<PhilosopherBackup> exportPhilosophers(long nrPhilosophers)
 			throws RemoteException {
 
 		Logging.log(Logger.Client, "Export " + nrPhilosophers
